cleaning dir after toppic

+++ src/base/file_util.cpp
@@ -147,10 +147,14 @@ void cleanDir(const std::string &path){
   for(fs::directory_iterator dir_iter(absolute(sp).parent_path()) ; dir_iter != end_iter ; ++dir_iter){
     std::string filename = dir_iter->path().string();
     if (filename.compare(0, base.length(), base) == 0){
-      if (filename != absolute(sp).string()&&!fs::is_directory(dir_iter->path())&&filename != base + ".OUTPUT_TABLE"){
+      if (filename != absolute(sp).string()&&!fs::is_directory(dir_iter->path())&&filename != base + ".OUTPUT_TABLE"&&filename != base + ".fasta"){
         fs::remove(dir_iter->path());
       }
-    }
+    } else if (filename.compare(filename.size() - 12, 12, "target_decoy") == 0){
+	  fs::remove(dir_iter->path());	
+	} else if (filename.compare(filename.size() - 6, 6, "target") == 0){
+	  fs::remove(dir_iter->path());	
+	}
   }
 }
